+++ src/lib/components/paginator/paginator.component.ts
@@ -1,3 +1,4 @@
+import { NgIf } from '@angular/common';
 import { Component, EventEmitter, Input, Output, inject } from '@angular/core';
 import { PaginableTablePagination } from '../../interfaces/paginable-table-pagination';
 import { PaginableService } from '../../services/paginable.service';
@@ -5,7 +6,9 @@ import { PaginableService } from '../../services/paginable.service';
 @Component({
 	selector: 'paginable-table-paginator',
 	templateUrl: './paginator.component.html',
-	styleUrls: ['./paginator.component.scss']
+	styleUrls: ['./paginator.component.scss'],
+	standalone: true,
+	imports: [NgIf]
 })
 export class PaginatorComponent {
 	private _configSvc = inject(PaginableService);