+++ src/lib/directives/paginable-table-not-found.directive.ts
@@ -1,12 +1,9 @@
 import { Directive, TemplateRef } from '@angular/core';
 
 @Directive({
-	selector: '[paginableTableNotFound]'
+	selector: '[paginableTableNotFound]',
+	standalone: true
 })
 export class PaginableTableNotFoundDirective {
-
-	constructor(
-		public template: TemplateRef<any>
-	) { }
-
+	constructor(public template: TemplateRef<any>) {}
 }